[connector] move decorators
